Split Harbor template tests for layout guard
